LLVM expects param3 of memset (which is an llvm intrinsic) to correspond to the size of an llvm pointer (4bytes). We had added modifications that make the C pointer 8 bytes. So change the paramter 3 of memset from size_t to int.

+++ system/lib/libc/musl/src/string/memset.c
@@ -1,7 +1,10 @@
 #include <string.h>
 #include <stdint.h>
 
-void *memset(void *dest, int c, size_t n)
+//LLVM expects param3 of memset (which is an llvm intrinsic) to correspond to the size of an llvm pointer (4bytes)
+//We had added modifications that make the C pointer 8 bytes. So change the sig of memset
+// void *memset(void *dest, int c, size_t n)
+void *memset(void *dest, int c, int n)
 {
 	unsigned char *s = dest;
 	size_t k;
